requirements.txt
asgiref==3.7.2
Brotli==1.1.0
colorama==0.4.6
dj-database-url==2.1.0
Django==4.2.5
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
gunicorn==21.2.0
mssql-django==1.3
packaging==23.2
Pillow==10.0.1
psycopg2-binary==2.9.9
PyJWT==1.7.1
pypng==0.20220715.0
pytz==2023.3.post1
qrcode==7.4.2
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
whitenoise==6.6.0
